mask_codec: fix int8_t overflow — cast masks to int32 before scaling
masks from extract_masks() are int8 dtype. Multiplying int8 * 63 overflows
(4 * 63 = 252 > 127 = int8 max). This blocked ALL AV1 encoding at 384x512.
Fix: cast to int32 before multiplication.

Test results (600 odd-frame masks at 384x512):
  AV1 CRF63: 59KB, 99.19% accuracy
  AV1 CRF50: 234KB, 99.77% accuracy
  H264 CRF30: 278KB, 99.90% accuracy
  Entropy: 583KB, 100% lossless

+++ src/tac/mask_codec.py
@@ -545,8 +545,10 @@ def encode_masks_monochrome(
     N, H, W = masks.shape
 
     # Scale class labels to byte range
+    # Cast to int32 first — masks may be int8 from SegNet extraction,
+    # and int8 * 63 overflows (4 * 63 = 252 > 127 = int8 max)
     scale_factor = 255 // (NUM_CLASSES - 1)
-    pixels = (masks * scale_factor).clamp(0, 255).to(torch.uint8).numpy()
+    pixels = (masks.to(torch.int32) * scale_factor).clamp(0, 255).to(torch.uint8).numpy()
 
     cmd = [
         "ffmpeg",
